Break up graphql module into separate submodules
