+++ server/src/graphql/edges/mod.rs
@@ -1,0 +1,5 @@
+mod person_squad;
+mod squad_member;
+
+pub use person_squad::*;
+pub use squad_member::*;
