+++ server/src/graphql/mod.rs
@@ -1,0 +1,35 @@
+pub mod edges;
+pub mod mutations;
+pub mod nodes;
+mod mutation_root;
+mod page_info;
+mod query_root;
+
+use crate::db;
+use crate::settings::Settings;
+use async_graphql::EmptySubscription;
+
+pub use mutation_root::*;
+pub use page_info::*;
+pub use query_root::*;
+
+pub type Schema = async_graphql::Schema<QueryRoot, MutationRoot, EmptySubscription>;
+
+pub fn make_schema(settings: Settings, pool: db::Pool) -> Schema {
+    let mut builder = Schema::build(QueryRoot {}, MutationRoot {}, EmptySubscription {})
+        .data(settings)
+        .data(pool)
+        .extension(|| async_graphql::extensions::Logger::default());
+
+    #[cfg(feature = "graphiql")]
+    {
+        builder = builder.extension(|| async_graphql::extensions::ApolloTracing::default());
+    }
+
+    #[cfg(not(feature = "graphiql"))]
+    {
+        builder = builder.disable_introspection();
+    }
+
+    builder.finish()
+}
