+++ server/src/graphql/mutations/mod.rs
@@ -1,0 +1,7 @@
+mod add_person_to_squad;
+mod new_person;
+mod new_squad;
+
+pub use add_person_to_squad::*;
+pub use new_person::*;
+pub use new_squad::*;
