+++ server/src/graphql/nodes/mod.rs
@@ -1,0 +1,11 @@
+mod person;
+mod squad;
+
+pub use person::*;
+pub use squad::*;
+
+#[async_graphql::Interface(field(name = "id", type = "String"))]
+pub enum Node {
+    Person(Person),
+    Squad(Squad),
+}
